rename internal tests to regular tests, I don't think that makes such a difference
